Call every knob by one name before the names set
The 1.0 naming-and-consistency pass, run across the whole public
surface: nine area audits over ~6,700 public declarations, ~150
renames applied, 918 files moved in step. The settled rules now live
in CLAUDE.md Conventions (one word per magnitude, make-prefixed
resource factories, short min/max, Int seeds, advance-by-seconds vs
step-by-counts, withExtension:, configuration Booleans as third-person
assertions); the judgment calls that still want a decision are the new
open-calls ledger in DESIGN-NOTES.

Wire formats hold: PhoneWire is positional binary, ParamStored's
rectangle case keeps its "rect" coding key, RemoteWire's new enums use
the old strings as raw values, and the video codec flag parses
case-insensitively. The tests caught one real regression the rename
machinery introduced (a shadowed loop variable turned a radial-basis
distance into zero) and the whole suite, the snapshot references, the
examples package, and all 473 Guide figures are green on the result.
Seven figures re-render because they draw API text that changed; two
more carry a new unstable mark, each proven by back-to-back renders.

+++ Guide/Figures/20-ParticleSimulations/ParticleLenia.swift
@@ -36,7 +36,7 @@ final class ParticleLeniaFigure: Sketch {
         noClear()
         for (i, panel) in [left, middle, right].enumerated() {
             let run = ParticleLenia(count: 1700, bounds: panel, spacing: 3.4,
-                                    seed: UInt64(21 + i))
+                                    seed: 21 + i)
             run.muG = settings[i].muG
             run.sigmaG = settings[i].sigmaG
             // One published step a frame, which is the unit the model's behavior is
